feat(slime_os/generation): add format v3 with u64 rights
C7.1 introduces generation format v3, widening the capability-grant
rights field to a little-endian u64 and the kernel/userspace rights
model (Rights = u64) and SpawnGrant ABI to match. Format v2 stays
decodable for the bounded rollback window: decode selects the grant
layout and authority-manifest hash width from the format version, so
retained v2 known-good generations still decode and their signed
releases still verify. The grandfathered RIGHT_MAP is renamed to the
object-specific RIGHT_BUFFER_MAP.

No new rights are gated yet; the SharedBufferFactory object that
consumes the freed bits lands in C7.2.

Verified: generation_check, contracts_check (with boot-contracts v2/v3
decode tests), transfer_check, test, fmt, lint, framework_safety_check.

+++ kernel/tests/spawn_authority.rs
@@ -45,7 +45,7 @@ fn panic(info: &core::panic::PanicInfo) -> ! {
     slime_os_kernel::test_panic_handler(info)
 }
 
-fn endpoint_cap(rights: u32) -> Capability {
+fn endpoint_cap(rights: u64) -> Capability {
     let (a, _b) = ipc::channel();
     Capability {
         object: KernelObject::Endpoint(a),
@@ -53,7 +53,7 @@ fn endpoint_cap(rights: u32) -> Capability {
     }
 }
 
-fn executable_cap(rights: u32) -> Capability {
+fn executable_cap(rights: u64) -> Capability {
     Capability {
         object: KernelObject::Executable {
             name: None,
@@ -144,7 +144,7 @@ fn table_rejects_rights_foreign_to_object_kind() {
     }
 }
 
-fn grant(slot: u32, rights: u32) -> task::SpawnGrant {
+fn grant(slot: u32, rights: u64) -> task::SpawnGrant {
     task::SpawnGrant { slot, rights }
 }
